backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth / security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12

# Config
pydantic-settings==2.6.1
pydantic[email]==2.10.3

# HTTP client (for Google token verification)
httpx==0.28.1

# Firebase / FCM
firebase-admin==6.6.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
aiosqlite==0.20.0

# Dev / lint (optional, helpful in container)
ruff==0.8.4
